Moved map from Main Page to Map page
- Also deleted some unused imports
